MachineLearning/requirements.txt
Flask>=2.0.0
numpy>=1.22.0
gevent>=23.9.1
pillow>=9.0.0
h5py>=3.1.0
tensorflow>=2.6.0
gunicorn>=20.0.0
flask-cors>=3.0.0
kaggle>=1.5.12
